requirements.txt
# Core ML
tensorflow-cpu==2.20.0
keras==3.10.0
numpy==1.26.4

# TensorFlow dependencies
protobuf==5.28.3
ml-dtypes==0.5.1
tensorboard==2.20.0

# Backend (FastAPI)
fastapi==0.115.0
uvicorn==0.30.6

# Database
sqlalchemy==2.0.23

# Image handling
pillow==10.4.0

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0
